permbench: add initial permbench command
This change adds the permbench command. This command will run a suite of
benchmarks against a targeted Perm service deployment. At the moment it
only contains a single benchmark but is written such that adding new
benchmarks is quick and familiar.

Please see the package documentation for more details about the
implementation.